Add Feign client for currency-service

See merge request xkollar3/online-banking-service!22

+++ transaction-service/src/main/java/cz/muni/fi/obs/data/dbo/TransactionDbo.java
@@ -9,7 +9,7 @@ import lombok.Builder;
 @Builder
 public record TransactionDbo(
 		@Id String id,
-		BigDecimal conversionRate,
+		Double conversionRate,
 		String withdrawsFrom,
 		String depositsTo,
 		BigDecimal withdrawAmount,
